rename: Cell 을 LabelCell 로 변경
